Move extension functions related to ClearlyDefined
The functionality to map ORT package coordinates to ClearlyDefined
coordinates is going to be needed by the scanner module as well.
Therefore, move the corresponding extension functions to the model
module. From there they can be accessed by both the scanner and
the analyzer modules.
